- minor cleanup, use proper datatypes, and Qt invokes the error itself no need for double check
svn path=/trunk/extragear/network/konversation/; revision=1002992

+++ src/dcc/transferrecv.cpp
@@ -663,7 +663,7 @@ namespace Konversation
         void TransferRecv::readData()                  // slot
         {
             //kDebug();
-            int actual = m_recvSocket->read( m_buffer, m_bufferSize );
+            qint64 actual = m_recvSocket->read( m_buffer, m_bufferSize );
             if ( actual > 0 )
             {
                 //actual is the size we read in, and is guaranteed to be less than m_bufferSize
